requirements.txt
streamlit>=1.46.0
pydantic-ai>=0.3.2
pydantic>=2.0
mammoth>=1.9.1 
python-dotenv>=1.0.0
PyPDF2>=3.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-anthropic>=0.1.0
langchain-groq>=0.1.0
python-docx>=1.1.0
pytesseract
pillow
requests